# Add Decimation/Interpolation to node class.