--- a/test/qa_settings.cpp
+++ b/test/qa_settings.cpp
@@ -203,17 +203,17 @@ const boost::ut::suite SettingsTests = [] {
         // define basic Sink->TestBlock->Sink flow graph
         auto &src = flow_graph.make_node<Source<float>>({ { "n_samples_max", n_samples } });
         expect(eq(src.n_samples_max, n_samples)) << "check map constructor";
-        expect(eq(src.settings().auto_update_parameters().size(), 4UL));
+        expect(eq(src.settings().auto_update_parameters().size(), 7UL));
         expect(eq(src.settings().auto_forward_parameters().size(), 1UL)); // sample_rate
         auto &block1 = flow_graph.make_node<TestBlock<float>>();
         auto &block2 = flow_graph.make_node<TestBlock<float>>();
         auto &sink   = flow_graph.make_node<Sink<float>>();
-        expect(eq(sink.settings().auto_update_parameters().size(), 5UL));
+        expect(eq(sink.settings().auto_update_parameters().size(), 8UL));
         expect(eq(sink.settings().auto_forward_parameters().size(), 1UL)); // sample_rate
 
         block1.context = "Test Context";
         block1.settings().update_active_parameters();
-        expect(eq(block1.settings().auto_update_parameters().size(), 6UL));
+        expect(eq(block1.settings().auto_update_parameters().size(), 9UL));
         expect(eq(block1.settings().auto_forward_parameters().size(), 2UL));
 
         expect(block1.settings().get("context").has_value());
@@ -226,7 +226,7 @@ const boost::ut::suite SettingsTests = [] {
         expect(block1.settings().get(keys1).empty());
         expect(block1.settings().get(keys2).empty());
         expect(block1.settings().get(keys3).empty());
-        expect(eq(block1.settings().get().size(), 7UL));
+        expect(eq(block1.settings().get().size(), 10UL));
 
         // set non-existent setting
         expect(not block1.settings().changed()) << "settings not changed";
@@ -276,7 +276,7 @@ const boost::ut::suite SettingsTests = [] {
         "empty"_test = [] {
             auto block = TestBlock<float>();
             block.init();
-            expect(eq(block.settings().get().size(), 7UL));
+            expect(eq(block.settings().get().size(), 10UL));
             expect(eq(std::get<float>(*block.settings().get("scaling_factor")), 1.f));
         };
 
@@ -287,7 +287,7 @@ const boost::ut::suite SettingsTests = [] {
             block.init();
             expect(eq(block.settings().staged_parameters().size(), 0u));
             block.settings().update_active_parameters();
-            expect(eq(block.settings().get().size(), 7UL));
+            expect(eq(block.settings().get().size(), 10UL));
             expect(eq(block.scaling_factor, 2.f));
             expect(eq(std::get<float>(*block.settings().get("scaling_factor")), 2.f));
         };
@@ -296,7 +296,7 @@ const boost::ut::suite SettingsTests = [] {
         "empty via graph"_test = [] {
             graph flow_graph;
             auto &block = flow_graph.make_node<TestBlock<float>>();
-            expect(eq(block.settings().get().size(), 7UL));
+            expect(eq(block.settings().get().size(), 10UL));
             expect(eq(block.scaling_factor, 1.f));
             expect(eq(std::get<float>(*block.settings().get("scaling_factor")), 1.f));
         };
@@ -304,7 +304,7 @@ const boost::ut::suite SettingsTests = [] {
         "with init parameter via graph"_test = [] {
             graph flow_graph;
             auto &block = flow_graph.make_node<TestBlock<float>>({ { "scaling_factor", 2.f } });
-            expect(eq(block.settings().get().size(), 7UL));
+            expect(eq(block.settings().get().size(), 10UL));
             expect(eq(block.scaling_factor, 2.f));
             expect(eq(std::get<float>(*block.settings().get("scaling_factor")), 2.f));
         };
@@ -314,7 +314,7 @@ const boost::ut::suite SettingsTests = [] {
         graph flow_graph;
         auto &block = flow_graph.make_node<TestBlock<float>>();
         block.settings().update_active_parameters();
-        expect(eq(block.settings().get().size(), 7UL));
+        expect(eq(block.settings().get().size(), 10UL));
 
         block.debug    = true;
         const auto val = block.settings().set({ { "vector_setting", std::vector{ 42.f, 2.f, 3.f } } });